+++ src/content/releases/2025-04-16.md
@@ -1,0 +1,22 @@
+---
+title: "Livelier Livestreams"
+description: "Serial now differentiates between standard videos and livestreamss, adding some nice UI and UX flair."
+publish_date: "2025-04-16"
+public: true
+---
+
+#### Features
+
+- Adds better livestream support
+  - Adds "Go Live" button that will bring you to the most recent part of a livestream
+  - Adds "Live" indicator for when you are caught up to a livestream
+
+#### Improvements
+
+- Adds a button to toggle windowed fullscreen
+  - Alternatively, toggle fullscreen with <kbd>`</kbd> or <kbd>f</kbd>
+- Adds buttons to zoom in and out the active video
+  - Alternatively, zoom in with <kbd>+</kbd> and out with <kbd>-</kbd>
+  - Minor adjustments to zoom, adding one more zoom in step
+- Cleaner logic for custom video player
+- Better shortcut logic when on video page
